Release 1.3.1
