+++ src/app/disease-form/common/formlyConfigs/disease-choice.ts
@@ -23,7 +23,7 @@ export function getDiseaseChoiceFields(diseaseOptions: DemisCoding[]): FormlyFie
     infoOutline,
     {
       template:
-        '<div class="info-link">Weiterführende Informationen zur Meldung gemäß § 6 IfSG finden Sie in der <a href="https://go.gematik.de/demis-info-disease" target="_blank">DEMIS-Wissensdatenbank</a></div>',
+        '<div class="info-link" id="knowledge-db-info-link">Weiterführende Informationen zur Meldung gemäß § 6 IfSG finden Sie in der <a href="https://go.gematik.de/demis-info-disease" target="_blank">DEMIS-Wissensdatenbank</a></div>',
     },
     {
       template: '<div class="question-title">Erkrankung *</div>',
