+++ src/app/disease-form/disease-form.component.scss
@@ -16,7 +16,7 @@
 
 :host ::ng-deep .mat-mdc-tab-label-container {
   width: 25vw;
-  background-color: rgb(249, 252, 252);
+  background-color: #FFFFFF;
   border-bottom-right-radius: 0;
   border-right-color: rgba(0, 3, 3, 0.12);
   border-right-style: solid;
@@ -74,7 +74,7 @@
 
 .nav-button {
   border-radius: 0;
-  background-color: white;
+  background-color: #FFFFFF;;
   color: #0f62fe;
   margin-right: 20px;
   border: 1px solid blue;
