+++ src/app/shared/components/autocomplete/autocomplete.component.scss
@@ -1,5 +1,5 @@
 .chip-list-wrapper {
-  background-color: white;
+  background-color: #FFFFFF;;
   margin-top: -1rem;
   margin-right: -1rem;
   margin-left: -1rem;
