+++ src/app/shared/components/paste-box/paste-box.component.scss
@@ -5,8 +5,6 @@
 
   button {
     padding: 8px 16px;
-    font-weight: 500;
-    font-family: 'Roboto', sans-serif;
     color: #0f62fe;
     border: 1px solid #0f62fe;
     border-radius: 0;
@@ -17,7 +15,7 @@
     cursor: pointer;
 
     &:hover {
-      background-color: #f1f4f4;
+      background-color: #F9F9F9;
     }
   }
